+++ include/mqt-core/fomac/FoMaC.hpp
@@ -24,7 +24,7 @@
 #include <utility>
 #include <vector>
 
-namespace qdmi {
+namespace fomac {
 /**
  * @brief Concept for ranges that are contiguous in memory and can be
  * constructed with a size.
@@ -178,7 +178,7 @@ inline auto throwIfError(int result, const std::string& msg) -> void {
  * @brief Class representing the FoMaC library.
  * @details This class provides methods to query available devices and
  * manage the QDMI session.
- * @note This class is a singleton
+ * @note This class is a singleton.
  * @see QDMI_Session
  */
 class FoMaC {
@@ -428,7 +428,7 @@ public:
           -> std::optional<bool>;
       /// @see QDMI_OPERATION_PROPERTY_SITES
       [[nodiscard]] auto getSites(const std::vector<Site>& sites = {},
-                                  const std::vector<double>& params = {})
+                                  const std::vector<double>& params = {}) const
           -> std::optional<std::vector<Site>>;
       /// @see QDMI_OPERATION_PROPERTY_MEANSHUTTLINGSPEED
       [[nodiscard]] auto
@@ -565,4 +565,4 @@ public:
   /// @see QDMI_SESSION_PROPERTY_DEVICES
   [[nodiscard]] static auto getDevices() -> std::vector<Device>;
 };
-} // namespace qdmi
+} // namespace fomac