+++ src/na/device/Generator.cpp
@@ -23,10 +23,12 @@
 #include <istream>
 #include <nlohmann/json.hpp>
 #include <ostream>
+#include <ranges>
 #include <spdlog/spdlog.h>
 #include <sstream>
 #include <stdexcept>
 #include <string>
+#include <tuple>
 #include <vector>
 
 namespace na {
@@ -224,6 +226,7 @@ auto writeSites(const Device& device, std::ostream& os) -> void {
 
     const std::vector limits{maxI, maxJ};
     std::vector indices{minI, minJ};
+    std::vector<std::tuple<size_t, int64_t, int64_t>> sites;
     for (bool loop = true; loop;
          loop = increment(indices, limits), ++subModuleCount) {
       // For every sublattice offset, add a site for repetition indices
@@ -235,9 +238,10 @@ auto writeSites(const Device& device, std::ostream& os) -> void {
         y += indices[0] * latticeVector1.y;
         x += indices[1] * latticeVector2.x;
         y += indices[1] * latticeVector2.y;
-        if (origin.x <= x && x < origin.x + extentWidth && origin.y <= y &&
-            y < origin.y + extentHeight) {
+        if (origin.x <= x && x <= origin.x + extentWidth && origin.y <= y &&
+            y <= origin.y + extentHeight) {
           // Only add the site if it is within the extent of the lattice
+          sites.emplace_back(id, x, y);
           os << ";\\\n  "
                 "var.emplace_back(MQT_NA_QDMI_Site_impl_d::makeUniqueSite("
              << id << "U, " << moduleCount << "U, " << subModuleCount << "U, "
@@ -251,6 +255,34 @@ auto writeSites(const Device& device, std::ostream& os) -> void {
                          static_cast<int64_t>(operation.region.size.height)) {
               os << ";\\\n  localOp" << operation.name
                  << "Sites.emplace_back(var.back().get())";
+            }
+          }
+          // this generator (same as the device implementation) only supports
+          // two-qubit local operations
+          for (const auto& operation : device.localMultiQubitOperations) {
+            if (operation.region.origin.x <= x &&
+                x <= operation.region.origin.x +
+                         static_cast<int64_t>(operation.region.size.width) &&
+                operation.region.origin.y <= y &&
+                y <= operation.region.origin.y +
+                         static_cast<int64_t>(operation.region.size.height)) {
+              for (const auto& [i2, x2, y2] :
+                   sites | std::views::take(sites.size() - 1)) {
+                if (operation.region.origin.x <= x2 &&
+                    x2 <=
+                        operation.region.origin.x +
+                            static_cast<int64_t>(operation.region.size.width) &&
+                    operation.region.origin.y <= y2 &&
+                    y2 <= operation.region.origin.y +
+                              static_cast<int64_t>(
+                                  operation.region.size.height) &&
+                    std::hypot(x2 - x, y2 - y) <=
+                        static_cast<double>(operation.interactionRadius)) {
+                  os << ";\\\n  localOp" << operation.name
+                     << "Sites.emplace_back(var.at(" << i2
+                     << ").get(), var.back().get())";
+                }
+              }
             }
           }
         }